# artifacts/api-server/src/routes/auth.ts
import { Router, type IRouter } from "express";
import { db, emailAccessLogTable } from "@workspace/db";
import { SubmitEmailBody, SubmitEmailResponse } from "@workspace/api-zod";

const ALLOWED_DOMAINS = ["lathropgpm.com", "kindredbravely.com"];

const router: IRouter = Router();

router.post("/auth/email", async (req, res): Promise<void> => {
  const parsed = SubmitEmailBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: "Invalid email address" });
    return;
  }

  const { email } = parsed.data;
  const domain = email.split("@")[1]?.toLowerCase();

  if (!domain || !ALLOWED_DOMAINS.includes(domain)) {
    res.status(403).json({
      error:
        "Access is restricted to authorized email domains (@lathropgpm.com or @kindredbravely.com)",
    });
    return;
  }

  await db.insert(emailAccessLogTable).values({ email });

  req.log.info({ email }, "Email access logged");

  res.json(SubmitEmailResponse.parse({ success: true }));
});

export default router;